Snake game
In the snake game I have created 4 methods: one for game setup, one to draw the field and the snake itself, one for the keyboard input, and one to control the movement of the snake. The snake can pass through the walls of the field, making the game more fun. It can only die, if it reaches his own tail, and that means that it's game over. Also there is a score, each time the snake eats a star (*) the score increases with 10.